Move version check logic into its own abstract presenter and its UI callbacks to its own view and inherit all of them accordingly
