+++ app/src/main/java/chat/rocket/android/authentication/login/presentation/LoginPresenter.kt
@@ -10,6 +10,7 @@ import chat.rocket.android.infrastructure.LocalRepository
 import chat.rocket.android.server.domain.*
 import chat.rocket.android.server.domain.model.Account
 import chat.rocket.android.server.infraestructure.RocketChatClientFactory
+import chat.rocket.android.server.presentation.CheckServerPresenter
 import chat.rocket.android.util.VersionInfo
 import chat.rocket.android.util.extensions.*
 import chat.rocket.common.RocketChatException
@@ -40,7 +41,8 @@ class LoginPresenter @Inject constructor(private val view: LoginView,
                                          settingsInteractor: GetSettingsInteractor,
                                          serverInteractor: GetCurrentServerInteractor,
                                          private val saveAccountInteractor: SaveAccountInteractor,
-                                         private val factory: RocketChatClientFactory) {
+                                         private val factory: RocketChatClientFactory)
+    : CheckServerPresenter(strategy, factory.create(serverInteractor.get()!!), view) {
     // TODO - we should validate the current server when opening the app, and have a nonnull get()
     private val currentServer = serverInteractor.get()!!
     private val client: RocketChatClient = factory.create(currentServer)
@@ -264,81 +266,5 @@ class LoginPresenter @Inject constructor(private val view: LoginView,
         val thumb = UrlHelper.getAvatarUrl(currentServer, me.username!!)
         val account = Account(currentServer, icon, logo, me.username!!, thumb)
         saveAccountInteractor.save(account)
-    }
-
-    private fun checkServerInfo() {
-        launchUI(strategy) {
-            val serverInfo = client.serverInfo()
-            val thisServerVersion = serverInfo.version
-            val isRequiredVersion = isRequiredServerVersion(thisServerVersion)
-            val isRecommendedVersion = isRecommendedServerVersion(thisServerVersion)
-            if (isRequiredVersion) {
-                if (isRecommendedVersion) {
-                    Timber.i("Your version is nice! (Requires: 0.62.0, Yours: $thisServerVersion)")
-                } else {
-                    view.alertNotRecommendedVersion()
-                }
-            } else {
-                if (!isRecommendedVersion) {
-                    view.blockAndAlertNotRequiredVersion()
-                    Timber.i("Oops. Looks like your server is out-of-date! Minimum server version required ${BuildConfig.REQUIRED_SERVER_VERSION}!")
-                }
-            }
-        }
-    }
-
-    private fun isRequiredServerVersion(version: String): Boolean {
-        return isMinimumVersion(version, getVersionDistilled(BuildConfig.REQUIRED_SERVER_VERSION))
-    }
-
-    private fun isRecommendedServerVersion(version: String): Boolean {
-        return isMinimumVersion(version, getVersionDistilled(BuildConfig.RECOMMENDED_SERVER_VERSION))
-    }
-
-    private fun isMinimumVersion(version: String, required: VersionInfo): Boolean {
-        val thisVersion = getVersionDistilled(version)
-        with(thisVersion) {
-            if (major < required.major) {
-                return false
-            } else if (major > required.major) {
-                return true
-            }
-            if (minor < required.minor) {
-                return false
-            } else if (minor > required.minor) {
-                return true
-            }
-            return update >= required.update
-        }
-    }
-
-    private fun getVersionDistilled(version: String): VersionInfo {
-        var split = version.split("-")
-        if (split.isEmpty()) {
-            return VersionInfo(0, 0, 0, null, "0.0.0")
-        }
-        val ver = split[0]
-        var release: String? = null
-        if (split.size > 1) {
-            release = split[1]
-        }
-        split = ver.split(".")
-        val major = getVersionNumber(split, 0)
-        val minor = getVersionNumber(split, 1)
-        val update = getVersionNumber(split, 2)
-        return VersionInfo(
-                major = major,
-                minor = minor,
-                update = update,
-                release = release,
-                full = version)
-    }
-
-    private fun getVersionNumber(split: List<String>, index: Int): Int {
-        return try {
-            split.getOrNull(index)?.toInt() ?: 0
-        } catch (ex: NumberFormatException) {
-            0
-        }
     }
 }
