+++ app/src/main/java/chat/rocket/android/authentication/login/presentation/LoginView.kt
@@ -1,10 +1,11 @@
 package chat.rocket.android.authentication.login.presentation
 
+import chat.rocket.android.authentication.server.presentation.VersionCheckView
 import chat.rocket.android.core.behaviours.InternetView
 import chat.rocket.android.core.behaviours.LoadingView
 import chat.rocket.android.core.behaviours.MessageView
 
-interface LoginView : LoadingView, MessageView, InternetView {
+interface LoginView : LoadingView, MessageView, InternetView, VersionCheckView {
 
     /**
      * Shows the form view (i.e the username/email and password fields) if it is enabled by the server settings.
@@ -186,14 +187,4 @@ interface LoginView : LoadingView, MessageView, InternetView {
      * Alerts the user about a wrong inputted password.
      */
     fun alertWrongPassword()
-
-    /**
-     * Alerts the user about the server version not meeting the recommended server version.
-     */
-    fun alertNotRecommendedVersion()
-
-    /**
-     * Block user to proceed and alert him due to server having an unsupported server version.
-     */
-    fun blockAndAlertNotRequiredVersion()
 }
