+++ app/src/main/java/chat/rocket/android/authentication/server/presentation/VersionCheckView.kt
@@ -1,0 +1,13 @@
+package chat.rocket.android.authentication.server.presentation
+
+interface VersionCheckView {
+    /**
+     * Alerts the user about the server version not meeting the recommended server version.
+     */
+    fun alertNotRecommendedVersion()
+
+    /**
+     * Block user to proceed and alert him due to server having an unsupported server version.
+     */
+    fun blockAndAlertNotRequiredVersion()
+}
