+++ app/src/main/java/chat/rocket/android/main/presentation/MainPresenter.kt
@@ -9,6 +9,7 @@ import chat.rocket.android.server.domain.*
 import chat.rocket.android.server.domain.model.Account
 import chat.rocket.android.server.infraestructure.ConnectionManagerFactory
 import chat.rocket.android.server.infraestructure.RocketChatClientFactory
+import chat.rocket.android.server.presentation.CheckServerPresenter
 import chat.rocket.android.util.extensions.launchUI
 import chat.rocket.android.util.extensions.registerPushToken
 import chat.rocket.common.RocketChatException
@@ -16,26 +17,25 @@ import chat.rocket.common.util.ifNull
 import chat.rocket.core.RocketChatClient
 import chat.rocket.core.internal.rest.logout
 import chat.rocket.core.internal.rest.me
-import chat.rocket.core.internal.rest.registerPushToken
 import chat.rocket.core.internal.rest.unregisterPushToken
 import timber.log.Timber
 import javax.inject.Inject
 
 class MainPresenter @Inject constructor(
-    private val view: MainView,
-    private val strategy: CancelStrategy,
-    private val navigator: MainNavigator,
-    private val tokenRepository: TokenRepository,
-    private val serverInteractor: GetCurrentServerInteractor,
-    private val localRepository: LocalRepository,
-    private val navHeaderMapper: NavHeaderViewModelMapper,
-    private val saveAccountInteractor: SaveAccountInteractor,
-    private val getAccountsInteractor: GetAccountsInteractor,
-    private val removeAccountInterector: RemoveAccountInterector,
-    private val factory: RocketChatClientFactory,
-    getSettingsInteractor: GetSettingsInteractor,
-    managerFactory: ConnectionManagerFactory
-) {
+        private val view: MainView,
+        private val strategy: CancelStrategy,
+        private val navigator: MainNavigator,
+        private val tokenRepository: TokenRepository,
+        private val serverInteractor: GetCurrentServerInteractor,
+        private val localRepository: LocalRepository,
+        private val navHeaderMapper: NavHeaderViewModelMapper,
+        private val saveAccountInteractor: SaveAccountInteractor,
+        private val getAccountsInteractor: GetAccountsInteractor,
+        private val removeAccountInterector: RemoveAccountInterector,
+        private val factory: RocketChatClientFactory,
+        getSettingsInteractor: GetSettingsInteractor,
+        managerFactory: ConnectionManagerFactory
+) : CheckServerPresenter(strategy, client = factory.create(serverInteractor.get()!!), view = view) {
     private val currentServer = serverInteractor.get()!!
     private val manager = managerFactory.create(currentServer)
     private val client: RocketChatClient = factory.create(currentServer)
@@ -48,10 +48,12 @@ class MainPresenter @Inject constructor(
     fun toSettings() = navigator.toSettings()
 
     fun loadCurrentInfo() {
+        checkServerInfo()
         launchUI(strategy) {
             try {
                 val me = client.me()
                 val model = navHeaderMapper.mapToViewModel(me)
+
                 saveAccount(model)
                 view.setupNavHeader(model, getAccountsInteractor.get())
             } catch (ex: Exception) {
